Delete  Exception

+++ src/main/java/com/example/WebService/Spring/Service/UserService.java
@@ -1,11 +1,15 @@
 package com.example.WebService.Spring.Service;
 
 import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessResourceFailureException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.example.WebService.Spring.Entites.User;
 import com.example.WebService.Spring.Repositories.UserRepository;
+import com.example.WebService.Spring.Service.Exceptions.DatabaseException;
 import com.example.WebService.Spring.Service.Exceptions.ResourceNotFoundException;
 
 @Service
@@ -27,9 +31,17 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void deleteUser (Long id) {
-		repository.deleteById(id);
-	}
+	public void deleteUser(Long id) {
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    }catch(DataIntegrityViolationException  e) {			
+	    	throw new DatabaseException(e.getMessage());
+		}
+	} 
 	
 	public User updateUser (Long id, User obj) {
 		User entity = repository.getReferenceById(id);
